LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/local-var-substitution-expected.txt


PASS Fallback directly in result
PASS Fallback via present, but invalid local
PASS Fallback is locally resolved (result)
PASS Fallback is locally resolved (local var)
